% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_example}
\alias{x_example}
\title{Example genotype matrix}
\format{
An integer matrix of dosages (0, 1, 2) with 20 samples in rows
(\code{Sample1}, \ldots, \code{Sample20}) and 10 SNPs in columns
(\code{SNP1}, \ldots, \code{SNP10}).
}
\source{
Simulated example data.
}
\usage{
x_example
}
\description{
A small simulated genotype dosage matrix used in the README and examples.
}
\keyword{datasets}
