% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GeneXCell}
\alias{GeneXCell}
\title{Cell-type marker gene reference}
\format{
A data frame containing cell-type marker genes.
}
\source{
Aran et al. (2017)
}
\usage{
GeneXCell
}
\description{
Marker-gene sets compiled from Aran et al. (2017).
}
\keyword{datasets}
