% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x_example}
\alias{x_example}
\title{Example genotype matrix}
\format{
A numeric matrix.
}
\source{
Included with the ctOWAS package.
}
\usage{
x_example
}
\description{
A small toy genotype matrix (samples x SNPs) used in the README.
}
\keyword{datasets}
